Added files for the project
Short codes here, based on the 4 first tasks that Loic asked me to do. Got to remember about Tkinter, actually, as it's a library i knew some years ago, and then first time coding some python via QGIS.

The next task is the 5th. You can see the tasks in the TODOLIST, which will be what i'll try to complete and code in the future.

On another note, the NDVI tif image is an image taken from a previous work from my school. Its only purpose was to test it out with my code, to get its dimension and the CRS the image is set in (task 4)