Updated code, comboBox changed
Using the classic comboBox instead of the QgsMapLayerComboBox for the following reason: Unable to change the order of the layers prompted by that comboBox. This is why i came back to use the classic comboBox, as it doesn't prompt vector layers, but only raster ones.

Next objectives will be to create a rectangle shape that covers the raster image.