Update of the code
Another test: Created a new shapefile, with some features; it isn't in the correct place where the image is, but it has the same size. Still need to work tho for the other images, as i need to get its coordinates of the most little point. Will also just work around rectangular images in the future